added listeners for live trading

+++ src/pages/Markets/Home/Parlay/components/TicketV2/TicketV2.tsx
@@ -82,6 +82,7 @@ import Tooltip from 'components/Tooltip';
 import { getTicketPayment, removeAll, setPaymentAmountToBuy } from 'redux/modules/ticket';
 import { SportsAmmData, TicketMarket } from 'types/markets';
 import { executeBiconomyTransaction, getGasFeesForTx } from 'utils/biconomy';
+import { getLiveTradingProcessorTransaction } from 'utils/liveTradingProcessor';
 import useSportsAmmDataQuery from '../../../../../../queries/markets/useSportsAmmDataQuery';
 import { getSportsAMMV2QuoteMethod, getSportsAMMV2Transaction } from '../../../../../../utils/sportsAmmV2';
 import ShareTicketModalV2 from '../ShareTicketModalV2';
@@ -134,6 +135,9 @@ const Ticket: React.FC<TicketProps> = ({ markets, setMarketsOutOfLiquidity }) =>
     const [openApprovalModal, setOpenApprovalModal] = useState(false);
     const [showShareTicketModal, setShowShareTicketModal] = useState(false);
     const [shareTicketModalData, setShareTicketModalData] = useState<ShareTicketModalProps | undefined>(undefined);
+
+    const [allowedLiveTrade, setAllowedLiveTrade] = useState(false);
+    const [processingLiveTrade, setProcessingLiveTrade] = useState(false);
 
     const [gas, setGas] = useState(0);
     const defaultCollateral = useMemo(() => getDefaultCollateral(networkId), [networkId]);
@@ -360,6 +364,59 @@ const Ticket: React.FC<TicketProps> = ({ markets, setMarketsOutOfLiquidity }) =>
         isDefaultCollateral,
     ]);
 
+    useEffect(() => {
+        const { sportsAMMV2Contract, liveTradingProcessorContract } = networkConnector;
+        if (sportsAMMV2Contract && liveTradingProcessorContract) {
+            const setLiveTradingListeners = async () => {
+                try {
+                    liveTradingProcessorContract.removeAllListeners('LiveTradeFullfilled');
+                    liveTradingProcessorContract.on(
+                        'LiveTradeFulfilled',
+                        (
+                            recipient,
+                            requestId,
+                            allow,
+                            gameId,
+                            sportId,
+                            typeId,
+                            position,
+                            buyInAmount,
+                            expectedPayout,
+                            collateral,
+                            timestamp
+                        ) => {
+                            if (recipient == walletAddress) {
+                                if (allow) {
+                                    setAllowedLiveTrade(true);
+                                    setProcessingLiveTrade(false);
+                                } else {
+                                    setAllowedLiveTrade(false);
+                                    setProcessingLiveTrade(false);
+                                }
+                            }
+                            console.log('LiveTradeFulfilled event triggered:', {
+                                recipient: recipient,
+                                requestId: requestId,
+                                allow: allow,
+                                gameId: gameId,
+                                sportId: sportId,
+                                typeId: typeId,
+                                position: position,
+                                buyInAmount: buyInAmount,
+                                expectedPayout: expectedPayout,
+                                collateral: collateral,
+                                timestamp: new Date(timestamp),
+                            });
+                        }
+                    );
+                } catch (e) {
+                    console.log(e);
+                }
+            };
+            setLiveTradingListeners();
+        }
+    }, [walletAddress]);
+
     const handleAllowance = async (approveAmount: BigNumber) => {
         const { sportsAMMV2Contract, sUSDContract, signer, multipleCollateral } = networkConnector;
         if (sportsAMMV2Contract && multipleCollateral && signer) {
@@ -425,25 +482,40 @@ const Ticket: React.FC<TicketProps> = ({ markets, setMarketsOutOfLiquidity }) =>
                 const expectedPayout = coinParser(totalBuyAmount.toString(), networkId);
                 const additionalSlippage = ethers.utils.parseEther('0.02');
 
-                const tx = await getSportsAMMV2Transaction(
-                    isVoucherSelected,
-                    overtimeVoucher ? overtimeVoucher.id : 0,
-                    collateralAddress,
-                    isDefaultCollateral,
-                    isEth,
-                    networkId,
-                    sportsAMMV2ContractWithSigner,
-                    sportsAMMV2ContractWithSigner,
-                    liveTradingProcessorContractWithSigner,
-                    // overtimeVoucherContractWithSigner,
-                    tradeData,
-                    usdPaid,
-                    collateralPaid,
-                    expectedPayout,
-                    referralId,
-                    additionalSlippage,
-                    isAA
-                );
+                let tx;
+                if (tradeData[0].live) {
+                    setProcessingLiveTrade(true);
+                    tx = await getLiveTradingProcessorTransaction(
+                        isVoucherSelected,
+                        collateralAddress,
+                        liveTradingProcessorContractWithSigner,
+                        tradeData,
+                        usdPaid,
+                        expectedPayout,
+                        referralId,
+                        additionalSlippage,
+                        isAA
+                    );
+                } else {
+                    tx = await getSportsAMMV2Transaction(
+                        isVoucherSelected,
+                        overtimeVoucher ? overtimeVoucher.id : 0,
+                        collateralAddress,
+                        isDefaultCollateral,
+                        isEth,
+                        networkId,
+                        sportsAMMV2ContractWithSigner,
+                        sportsAMMV2ContractWithSigner,
+                        // overtimeVoucherContractWithSigner,
+                        tradeData,
+                        usdPaid,
+                        collateralPaid,
+                        expectedPayout,
+                        referralId,
+                        additionalSlippage,
+                        isAA
+                    );
+                }
 
                 const txResult = isAA ? tx : await tx.wait();
 
@@ -455,11 +527,27 @@ const Ticket: React.FC<TicketProps> = ({ markets, setMarketsOutOfLiquidity }) =>
                             networkId,
                         },
                     });
-                    refetchBalances(walletAddress, networkId);
-                    toast.update(id, getSuccessToastOptions(t('market.toast-message.buy-success')));
-                    setIsBuying(false);
-                    setCollateralAmount('');
-                    dispatch(removeAll());
+                    if (tradeData[0].live) {
+                        while (processingLiveTrade) {}
+                        if (allowedLiveTrade) {
+                            refetchBalances(walletAddress, networkId);
+                            toast.update(id, getSuccessToastOptions(t('market.toast-message.buy-success')));
+                            setIsBuying(false);
+                            setCollateralAmount('');
+                            dispatch(removeAll());
+                            setAllowedLiveTrade(false);
+                        } else {
+                            setIsBuying(false);
+                            refetchBalances(walletAddress, networkId);
+                            toast.update(id, getErrorToastOptions(t('common.errors.unknown-error-try-again')));
+                        }
+                    } else {
+                        refetchBalances(walletAddress, networkId);
+                        toast.update(id, getSuccessToastOptions(t('market.toast-message.buy-success')));
+                        setIsBuying(false);
+                        setCollateralAmount('');
+                        dispatch(removeAll());
+                    }
                 }
             } catch (e) {
                 setIsBuying(false);
